Solve and enforce ruleset on errormessages
Ruff rules EM101 and EM102.

## What it does
Checks for the use of string literals in exception constructors.

## Why is this bad?
Python includes the  in the default traceback (and formatters
like Rich and IPython do too).

By using a string literal, the error message will be duplicated in the
traceback, which can make the traceback less readable.